Move the following files to common in order to use in other packages: - ValidationResult - SaveAuthenticationTokenResult - ValidateEmailUseCase - ValidatePasswordUseCase - SaveAuthenticationTokenUseCase
