+++ app/src/test/java/com/davidread/habittracker/common/usecase/ValidatePasswordUseCaseTest.kt
@@ -1,6 +1,6 @@
-package com.davidread.habittracker.signup.usecase
+package com.davidread.habittracker.common.usecase
 
-import com.davidread.habittracker.signup.model.ValidationResult
+import com.davidread.habittracker.common.model.ValidationResult
 import org.junit.Assert
 import org.junit.Test
