+++ app/src/test/java/com/davidread/habittracker/signup/viewmodel/ClickSignUpButtonTest.kt
@@ -9,7 +9,7 @@ import com.davidread.habittracker.signup.model.SignUpTextFieldViewState
 import com.davidread.habittracker.signup.model.SignUpViewEffect
 import com.davidread.habittracker.signup.model.SignUpViewIntent
 import com.davidread.habittracker.signup.model.SignUpViewState
-import com.davidread.habittracker.signup.model.ValidationResult
+import com.davidread.habittracker.common.model.ValidationResult
 import com.davidread.habittracker.signup.usecase.SignUpFlowUseCase
 import com.davidread.habittracker.testutil.MainDispatcherRule
 import io.mockk.clearAllMocks
